+++ vite/src/routes/extensions/components/extension-card.tsx
