+++ vite/src/routes/extensions/components/extension-details.tsx
